Move all mqtt mapping related stuff into connector.py
